voice_clone/requirements.txt
streamlit==1.32.0
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
groq==0.4.2
python-multipart==0.0.9
PyPDF2==3.0.1
TTS==0.22.0
numpy==1.22.0  # Downgraded to match TTS's requirement
soundfile==0.12.1
